requirements.txt
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.7.0
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.5.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
dlib==19.24.6
dnspython==2.1.0
duplicity==0.8.21
entrypoints==0.4
evdev==1.6.1
fasteners==0.14.1
filelock==3.13.4
Flask==3.0.3
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
gdown==5.2.0
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gpg==1.16.0
gspread==6.1.4
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==2.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.3
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
natsort==8.0.2
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.1
notebook==6.4.8
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==21.3
pandas==2.1.1
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.2.0
plotly==5.24.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
proto-plus==1.24.0
protobuf==5.28.2
ptyprocess==0.7.0
PuLP==2.7.0
py==1.10.0
py2flowchart==0.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.7
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-xlib==0.33
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rich==13.7.1
rsa==4.9
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
six==1.16.0
soupsieve==2.3.1
ssh-import-id==5.11
sympy==1.9
systemd-python==234
tenacity==9.0.0
terminado==0.13.1
testpath==0.5.0
tornado==6.1
tqdm==4.66.5
traitlets==5.1.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.25.3
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==3.0.2
widgetsnbextension==2.0.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
zipp==1.0.0
